Moved `as_ascii` and `indent` to new `utils.py`
